Move order mappers to details package for better organization
Moved WooPosOrderItemMapper and WooPosOrderStatusMapper from orders
package to orders.details package. These mappers handle order
presentation/display concerns and fit better with other display logic.
