+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/orders/details/WooPosOrderStatusMapper.kt
@@ -1,7 +1,9 @@
-package com.woocommerce.android.ui.woopos.orders
+package com.woocommerce.android.ui.woopos.orders.details
 
 import com.woocommerce.android.R
 import com.woocommerce.android.model.Order
+import com.woocommerce.android.ui.woopos.orders.OrderStatusColorKey
+import com.woocommerce.android.ui.woopos.orders.PosOrderStatus
 import com.woocommerce.android.viewmodel.ResourceProvider
 import java.util.Locale
 import javax.inject.Inject
